requirements_ml.txt
certifi==2023.7.22
charset-normalizer==3.3.0
cmdstanpy==1.2.0
contourpy==1.1.1
convertdate==2.4.0
cycler==0.12.0
ephem==4.1.4
fonttools==4.43.0
holidays==0.34
idna==3.4
importlib-resources==6.1.0
joblib==1.3.2
kiwisolver==1.4.5
lightgbm==4.1.0
LunarCalendar==0.0.9
matplotlib==3.8.0
numpy==1.26.0
packaging==23.2
pandas==2.1.1
patsy==0.5.3
phik==0.12.3
Pillow==10.0.1
prophet==1.1.4
PyMeeus==0.5.12
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
stanio==0.3.0
statsmodels==0.14.0
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.6
